Revise apply_batch -> apply_blob.
This commit makes a significant change to the core STF API.
The methods apply_batch and apply_proof are combined into a single
method "apply_blob". This method takes an iterator of bytes rather
than an associated type as an argument, and returns a BatchReceipt
associated type, which itself contains a vector of TransactionReceipts.

This allows the node implementation to be much more generic, while
leaving full expressivity to STF developers.

Update the AppTemplate to use the new trait. Finally, fix the imports
in the sequencer package to avoid importing rocksdb when built without
default features.