Cleanup. Introduce OrderedReadsAndWrites type.
This commit removes StorageInternalCache from the API of
Storage::validate_and_commit and replaces it with a new
type `OrderedReadsAndWrites`. The new type is both more secure
(it doesn't rely on the order of iterators on the HashMap type for
correctness) and surfaces less implementation details.

This commit also simplifies the implementation of
RevertableDelta, removing its cachelog in favor of a simple
hashmap of writes. This makes RevertableDelta much easier to
reason about, and removes the need for complicated merging logic.