Prover demo (#229)
* Update resolvers. Feature gate demo-app deps

* Revise apply_batch -> apply_blob.

This commit makes a significant change to the core STF API.
The methods apply_batch and apply_proof are combined into a single
method "apply_blob". This method takes an iterator of bytes rather
than an associated type as an argument, and returns a BatchReceipt
associated type, which itself contains a vector of TransactionReceipts.

This allows the node implementation to be much more generic, while
leaving full expressivity to STF developers.

Update the AppTemplate to use the new trait. Finally, fix the imports
in the sequencer package to avoid importing rocksdb when built without
default features.

* Add Witness Type to trait (#227)

* Start working on trait modification

* Fix after rebase

* Clean ups and types fixing

* Settle on changes

* Clean up comments

* Cleaning up

* Remove uneeded derives

* Rename StorageSpec to MerkleProofSpec

* Simplify revertable delta. Bug fixes in apply_blob

Currently, apply_blob sometimes destroys the working_set
when exiting early. Save the working set on all code paths.

* Introduce OrderedReadsAndWrites type.

This commit removes StorageInternalCache from the API of
Storage::validate_and_commit and replaces it with a new
type `OrderedReadsAndWrites`. The new type is both more secure
(it doesn't rely on the order of iterators on the HashMap type for
correctness) and surfaces less implementation details.

This commit also simplifies the implementation of
RevertableDelta, removing its cachelog in favor of a simple
hashmap of writes. This makes RevertableDelta much easier to
reason about, and removes the need for complicated merging logic.

* Rename *apply_batch to apply_blob

* Add explanation for why blob is impl

* Use matches! instead of if let where appropriate

* Use new storage constructor in app template

---------

Co-authored-by: Preston Evans <preston.b.evans@gmail.com>