feat: update `StateMap` to take borrowee of keys
This commit updates the API of `StateMap` to behave as common map
implementations in Rust: the argument of a get method will receive a
value that can be borrowed from the concrete key type.

This is useful for recurrent situations such as when we want to perform
a get from a map of `Vec<u8>` without cloning the key index, by using a
slice.

The focus of the PR is to touch the public API of the map and trait,
without changing too much the internals. A new issue is introduced to
propose an upgrade: #643

It also updates the public interface of the storage to take references
by default, instead of owned values. It will allow the user to perform
storage queries without necessarily cloning the values.

Resolves #427