feat: add sovereign.toml manifest file
This commit introduces a manifest file that can be parsed by derive
macros and have its contents read at compile-time.

It will read recursively the directory of the call site of the
macro until it either finds a `sovereign.toml` file, or the path is
depleted.

The implementation for the source file location of proc macros depends
on `procmacro2_semver_exempt`, and it will fallback to the manifest path
of the `sov-modules-macros` library. The fallback may contain multiple
edge cases as cargo will cache the downloaded crate into a
`$HOME/.cargo` dir, making a recursion tree impossible to reach the
workspace under normal circumstances.

The aforementioned problem will be handled on the issue https://github.com/Sovereign-Labs/sovereign-sdk/issues/786