feat: update sovereign.toml resoultion to relative path (#855)
* feat: update sovereign.toml resoultion to relative path

Prior to this commit, the resolution of the sovereign.toml dependend on
unstable features of proc-macro2, available only for recent toolchain
versions.

This commit introduces the path resolution of the manifest file to
recurse from the target directory. It will allow endpoints to set their
own manifest file, replacing the ones bundled with the modules.

It will also read an optional `SOVEREIGN_MANIFEST` that, if set, will
override the manifest path. The manifest path is expected to be resolved
at compile time, making it possible to construct const/static structures
with it.

* remove deprecated .env