feat: downgrade WorkingSet to sov-modules-api (#831)
* feat: downgrade WorkingSet to sov-modules-api

`WorkingSet`, defined in `sov-state`, might have definitions specified
under `Context`, that is defined in `sov-modules-api`.

Given `sov-state` is a dependency of `sov-modules-api`, this means
`WorkingSet` must be moved to the latter; otherwise, such relationship
will not be possible.

One example is the gas meter implementation that requires runtime gas
price information that is best defined under `Context`. The gas will be
deducted from the `WorkingSet`, but must be computed via the `Context`,
making this relationship overly convoluted, unless `WorkingSet` contains
a generic `Context`.

This commit solely moves the `WorkingSet` to the `sov-modules-api`,
without making changes on its structures. A subsequent commit will
introduce `Context` as generic attribute of `WorkingSet`.

* revert modules change onto sov-modules-api

* promote map proof_* fns to attester

* fix merge remainder nits

* fix remainder nit

* fix more nits

* fix ci nit

* fix ci nit

* fix remainder ci nits

* fix ci expected stderr msgs

* fix CI nit

* fix doc nits

* fix doctest nit