feat: add arbitrary feature to sov-db (#853)
This commit introduces arbitrary derive for the types defined in sov-db.

It also introduces a fallible wrapper for proptest, that expects the
types to be infallible. However, the database is created on disk; given
the fallible nature of I/O, they must also be subject to fail.

OTOH, a disk write failure is not a logical bug of the SDK, so maybe it
would be ok to panic. This could however be inconvenient, as any fuzzing
process would be interrupted by such panic.