feat: add addressable trait
Prior to this commit, the module info was the sole responsible to
provide a module address with its dependencies.

If the module info is ever to increase in complexity, such complexity
will be leaked to functionality that expects to handle only its
addresses, such as sort modules per deps/check cyclic dependencies.

This commit introduces `Addressable`, a trait that will be implemented
for every `ModuleInfo` and will provide the addresses, when required.