fix: broken fuzz lib (#984)
* fix: broken fuzz lib

Prior to this commit, there was no CI check for the fuzz library.

Since it is isolated from the workspace (so it won't be affected by fuzz
specific build configuration), it could be silently broken by a breaking
change.

This commit introduces a fix to `AccountConfig`, and adds the fuzz check
to the link job of the `Makefile`.

* Suggested CI cache fix

Signed-off-by: Filippo Costa <filippo@sovlabs.io>

---------

Signed-off-by: Filippo Costa <filippo@sovlabs.io>
Co-authored-by: Filippo Costa <filippo@sovlabs.io>