feat: add `no-std` to `rollup-interface`
This commit introduces a `no-std` build for `rollup-interface`. This is
required to create WASM modules for module implementations, having one
use-case the web wallet.

We follow a similar approach to `borsh` and create a `maybestd` module
so dependencies can rely on that for fallback implementations of common
stdlib components that are available either on `alloc` or commonly used
crates.

Most of the requirements exists on `alloc`, but we do have a couple of
exceptions. The first is `HashMap`, that we fallback for `hashbrown`
when `std` isn't available. We shouldn't expect issues from risc0
integration as `risc0` code will be compiled with `std`.

The second is `Mutex`; for that, we fallback to `spin`, another common
choice on the Rust ecosystem.

Finally, this commit introduces a couple of changes on the root
`Cargo.toml` to not use default-features for crates that offers `no-std`
support.