chore: freeze clap version
Clap introduced a change that breaks as it requires rustc to be 1.66,
and stable might be behind that.

This might be a violation of SEMVER, but it is debatable, as the rust
guidelines suggests compiler updates to be introduced as patch - which
is a rather questionable decision as this might break.

This commit freezes the clap version so it will be updated manually,
when required. With that, we don't incur risks of breaking changes on
patch update.

Reference issue:
https://github.com/clap-rs/clap/issues/5158