feat: add `no-std` to `rollup-interface` (#1077)
* feat: add `no-std` to `rollup-interface`

This commit introduces a `no-std` build for `rollup-interface`. This is
required to create WASM modules for module implementations, having one
use-case the web wallet.

We follow a similar approach to `borsh` and create a `maybestd` module
so dependencies can rely on that for fallback implementations of common
stdlib components that are available either on `alloc` or commonly used
crates.

Most of the requirements exists on `alloc`, but we do have a couple of
exceptions. The first is `HashMap`, that we fallback for `hashbrown`
when `std` isn't available. We shouldn't expect issues from risc0
integration as `risc0` code will be compiled with `std`.

The second is `Mutex`; for that, we fallback to `spin`, another common
choice on the Rust ecosystem.

Finally, this commit introduces a couple of changes on the root
`Cargo.toml` to not use default-features for crates that offers `no-std`
support.

* update fmt

* add to_string to mock da test

* fix sov-schema-db

* add `no-std` to `sov-schema-db`

* restore test-data

* fix format on sov-schema-db

* add no-std directive to sov-schema-db

* add `no-std` to `sov-db`

* add tokio to sov-stf-runner/sov-db

* add default-features to byteorder

* add `tokio` as default feature of `sov-db`

* restore nightly json files

* restore more json files

* move std feature shield bloat code to dedicated sub-modules

* move bloat feature shield to dedicated use_std modules

* Update rollup-interface/src/state_machine/mocks/zk_vm.rs

Co-authored-by: Filippo Neysofu Costa <filippo@sovlabs.io>

* revert sov-db & sov-schema-db

* restore split impl for std bincode

* reorg sov-schema-db manifest w/ default-features

* move da-mock-service to its own module

* upd cargo lock

* remote stall file

* add no-std CI check

* fix cargo docs

* require borsh features for sov-db

* add check_no_std to required all-green ci jobs

* clean check_no_std ci cache commands

---------

Co-authored-by: Filippo Neysofu Costa <filippo@sovlabs.io>