feat: add sov-modules-core
This commit introduces `sov-modules-core`, a new workspace member that
will contain the required declarations for `no-std` `DispatchCall`
runtime message declarations.

The new module is re-exported via `sov-state` and `sov-modules-api` to
minimize breaking changes downstream.

Finally, it introduces a `RefCount` in `sov-rollup-interface` that will
contain atomic synchronization, if the compile target contains atomic
pointers. It safely fallback to `Rc`, that is a reference count without
atomic synchronization, but equivalent API, that will be the used type
on targets without synchronization. This can replace all the `Arc`
instances in the future, but for now, it is required only for the
previously `sov-state` storage/cache keys and values, so the `Storage`
trait can be properly defined.