feat: add try_from slice for signature & keys (#1145)
This commit introduces a trait bound requirement for PublicKey and
Signature that will allow it to be recreated from bytes.

Prior to this commit, such implementation existed only for `FromStr` in
native. However, in some cases - such as in the wallet - the public keys
will be recreated from raw bytes computed from 3rd party libraries. The
serialization of such crypto primitives is a standard, so they can
safely interop with our implementations.