refactor: simplify CliWalletArg decodable message
This commit removes the `CliWallet` and `CliFrontEnd` traits.
Additionally, it refactors the `CliWalletArg` to encode runtime call
messages directly into a readable format, thereby eradicating an
intermediary representation.

The prior implementation employed an intermediary representation that
was adjustable by users but necessitated manual conversion into a
decodable message format. As of now, this step is deemed superfluous
because the readable transaction can embody the new `CliWalletArg` and
offer decoding functionality instantaneously. Typically, the input will
be a string that can be translated from a common serialization format
like JSON or YAML. By default, JSON is utilized; however, other common
serialization formats can also be supported.

The changeset aims to make frontend portrayal of superstructures
possible for runtime call messages. It simplifies the decoding process
so that it can be seamlessly plugged into wrapper superstructures
containing additional metadata. One potential instance is
`UnsignedTransaction` as a wrapping superstructure for runtime call
data, with relevant protocol-level information pertaining to the rollup.

Moreover, this commit renames `sov_modules_api::cli::FileNameArg` to
`JsonFileNameArg`. This type inherently represents the path to a JSON
file; hence, the name must reflect that. This allows for effortless
addition of paths symbolizing other formats in the future (e.g.,
`YamlFileNameArg`).