feat: add `UnsignedTransaction`
This commit introduces `UnsignedTransaction`, an intermediary type that
will contain transaction metadata before it is signed.

To maintain consistency with the JSON runtime call messages serialized
format, we preserve this format because the serialization occurs prior
to the unsigned transaction serialization. The runtime will define a
module name for its implementation; therefore, the runtime call will
have a custom field name, so it must be serialized beforehand.

In order to maintain simplicity, two CLI arguments are added to sov-cli
to contain the two new scalar attributes of the transaction: chain id
and gas tip.

The chain ID is introduced to prevent cross-chain transaction
reproduction. The signed transaction will also take the chain id field
as part of the signature; therefore, a signature is valid only for a
specified chain.

The gas tip is an incentive for the sequencer to pick the given
transaction from the pool.