refactor: use cargo ws root for constants.json path
This commit enhances the functionality of `constants.json` path
resolution by employing `cargo metadata` output to establish the
manifest path of the workspace root in the current working directory. As
a result, it is now feasible to determine the absolute path of the
constants file without recourse to a temporary build script hack.

Additionally, this commit eradicates the aforementioned hack that wrote
data into the git working set tree.

Resolves #1204