feat: add gas limit to tx
With this commit, we implement a gas limit for transactions. The
introduced value signifies the available gas resources for a transaction
prior to its execution by the runtime. Upon transaction dispatch
finalization, the bank module will consume the excess gas and distribute
the tip from the sender's account.

To identify the sender address, an event `accounts::sender` has been
established between accounts and the bank. This is essential because the
bank cannot determine or access context while executing hooks.

The definition of the dynamic gas price elasticity remains undetermined
with this commit, as it relies on chain state and execution history for
calculation. Furthermore, this change does not impact the accrued gas
rewards held by the sequencer, which will be transferred to his bank
account upon block execution completion.