feat: add associated types as tx hooks args
This commit introduces associated types to the `TxHooks` trait,
enhancing its ability to define custom interactions between the trait
and its implementing modules. The associated types allow for greater
flexibility in the implementation of `TxHooks`, enabling diverse use
cases beyond the current `Accounts` module.

The current `Accounts` implementation of `TxHooks` serves the purpose of
extracting a public key from a transaction and mapping it to a rollup
address, creating an execution context. This functionality is achieved
through the `pre_dispatch_tx_hook` signature which returns a
`C::Address`. However, this tight coupling with the `accounts` module
restricts future expansions of `TxHooks`.

With this modification, hooks gain increased versatility and modules are
free to implement them as per their requirements. Additionally, runtimes
can consume these implementations, utilizing them as essential building
blocks, while the banks module is empowered to enforce a gas cap on the
pre-dispatch hook.