feat: add gas total to tx receipt
This commit introduces the "gas total" as a new attribute in the
transaction receipt. With this addition, the gas price elasticity can be
calculated based on the demand of previous blocks.

This feature also allows setting a base fee for every transaction, which
is paid by the sender to cover processing costs such as deserialization
and signature verification. Insufficient funds from the sender will
cause the fee to be charged to the sequencer, leading to a failed
transaction and thwarting potential attacks. The sequencer should have
sufficient funds available from their staked amount to cover these
expenses; the implementation of the sequencer registry is responsible
for ensuring the sequencer possesses adequate funds for selecting
transactions to assemble into a blob.

To define the base cost constant, an adjustment was made to make this
feature available under the `sov-modules-macros`. Due to the need for
defining slices of unknown size for the constants file, this
functionality had to be provided there as well.