Fix UniversalWallet serialization for web3 SDK (#1966)
Add UniversalWallet overrides at the type level for every implementer of CryptoSpec, rather than overriding it at the field level inside Transaction.
This ensures the correct byte array length is used, whereas in the Transaction definition, it was not possible to use a generic length.
Add a ByteDisplayable bound which should ensure the override is to a bytevec or bytearray, and allows continuing to specify "display = hex" in the attribute.