Fix queue behaviour on sequencer resync.
The queue assumes that all transactions go through it. On resync, this isn't true, and the API state will increase independently of the queue. We detect this now by a) checking the original_tx_queue_id and b) detecting SequencerNotReady errors; both of these cases invalidate the entire nonce queue and cause a wipe.