Check all resync conditions before recovery in update_state.
Recovery should only trigger once we are at the chain tip - it's incorrect to use old data to decide whether to recover or not. This was mostly already the case but `condition_node_is_unsynced_and_doesnt_know_it` was being checked after the recovery conditions. This commit moved the match branch for this up so that if the condition is true we avoid triggering recovery based on stale information.