Replace nonce queue with single-threaded implementation (#2167)
* Inc gas_limit

* Timing oracle

* Extract metrics tracking

* Simplify handlers

* Create struct to dedupe handlers where clauses

* Move signing methods into Handlers struct

* Add eth_sendRawTransactionSync

* Extract receipt getting

* Fix foundry

* Use RpcResult

* Implement tx timeout on eth_send_raw_transaction_sync

* Don't swallow parsing errors

* Fix timeout message

* revert nomt

* Switch to nomt without page elision

* Fill zero hash for pending receipts

* Add extra information when the sequencer nonce queues reject a tx

* fmt

* lockfile needed changing for some reason

* Improve invalid nonce error message

* Fixes a bug in sequencer nonce queue when a transaction is rejected. (#2131)

Fixes a bug in sequencer nonce queue when a transaction is rejected.
The nonce queue optimistically keeps track of nonces submitted for execution, since the actual API state only updates after a transaction is persisted to the DB; tracking the executing nonce locally allows queueing up transactions with consecutive nonces without waiting for the DB latency.
The problem was that this locally tracked "last submitted for execution" nonce did not get reset if said submitted transaction actually got rejected. This had two consequences: first and less importantly, transactions would not be evicted even when no longer having all pre-requisites. This could've had a theoretical memory impact if someone filled their queue with transactions and then sent a rejected transaction at the head, resulting in the sequencer keeping the entire queue in memory forever. The more important issue was that the sequencer would then consider transactions with the next nonce valid and would submit them for execution. They would then always get rejected due to the user's nonce, of course, not having been actually incremented by the previous (rejected) transaction. This would spiral into a vicious cycle of invalidating and rejecting large swathes of transactions and causing terrible user experience.

* Bump sequencer queue timeout, sendrawtransaction_sync timeout, and add a debug log to all ethereum transactions ingested (#2136)

* Initial implementation. Integration tests pass, needs refactoring and unit tests

* Fix queue cleanup. Small non-persisted adjustment. Initial unit test porting.

* Remove non-working unit tests for now.

* Remove panics from queue for now

* Revert "Remove non-working unit tests for now."

This reverts commit 18757f2e67c8bb05d3b085b803097a419c0af78a.

* Refactor and convert most unit tests to new implementation

* Replacement unit tests added

* PR comments: disable pre-req checks on timeout, add debug assertions

* fmt and prints

* Squash single-threaded nonce queue impl

* Fix queue behaviour on sequencer resync.

The queue assumes that all transactions go through it. On resync, this isn't true, and the API state will increase independently of the queue. We detect this now by a) checking the original_tx_queue_id and b) detecting SequencerNotReady errors; both of these cases invalidate the entire nonce queue and cause a wipe.

* Allow reverted txs onchain via config (#2141)

---------

Co-authored-by: Theodore Bugnet <theodore@sovlabs.io>

* Fix queue behaviour on sequencer resync.

The queue assumes that all transactions go through it. On resync, this isn't true, and the API state will increase independently of the queue. We detect this now by a) checking the original_tx_queue_id and b) detecting SequencerNotReady errors; both of these cases invalidate the entire nonce queue and cause a wipe.

* PR comments

* Explicitly reject transactions on queue wipe, and restore unknown reason error when oneshot is dropped

* Fix spurious config changes

* Fix types for latest dev

* fmt

* fix tests

---------

Co-authored-by: bkolad <blazejkolad@gmail.com>
Co-authored-by: Leonid Logvinov <logvinov.leon@gmail.com>
Co-authored-by: Preston Evans <preston.b.evans@gmail.com>
Co-authored-by: Preston Evans <32944016+preston-evans98@users.noreply.github.com>